Moved credential file editing to the Plugin Management tab
