+++ src/main/java/net/pms/messages.properties
@@ -216,8 +216,8 @@ NetworkTab.50=Installing
 NetworkTab.51=Edit UMS configuration file manually
 NetworkTab.52=Error saving config file
 NetworkTab.53=Description
-NetworkTab.54=Edit UMS credential file
-NetworkTab.55=Error saving credential file
+NetworkTab.54=Edit Plugin Credential File
+NetworkTab.55=Error saving plugin credential file
 PluginTab.0=Installed Plugins
 PluginTab.1=Available Plugins
 PMS.0=No renderers were found
