+++ src/main/java/net/pms/newgui/PluginTab.java
@@ -4,6 +4,7 @@ import com.jgoodies.forms.builder.PanelBuilder;
 import com.jgoodies.forms.factories.Borders;
 import com.jgoodies.forms.layout.CellConstraints;
 import com.jgoodies.forms.layout.FormLayout;
+import java.awt.BorderLayout;
 import java.awt.Component;
 import java.awt.ComponentOrientation;
 import java.awt.Font;
@@ -11,6 +12,11 @@ import java.awt.GridLayout;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.MouseEvent;
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.Locale;
 import javax.swing.JButton;
@@ -22,6 +28,7 @@ import javax.swing.JPanel;
 import javax.swing.JProgressBar;
 import javax.swing.JScrollPane;
 import javax.swing.JTable;
+import javax.swing.JTextArea;
 import javax.swing.SwingUtilities;
 import javax.swing.table.TableColumn;
 import net.pms.Messages;
@@ -59,6 +66,84 @@ public class PluginTab {
 		JComponent cmp = builder.addSeparator(Messages.getString("NetworkTab.5"), FormLayoutUtil.flip(cc.xyw(1, 1, 9), colSpec, orientation));
 		cmp = (JComponent) cmp.getComponent(0);
 		cmp.setFont(cmp.getFont().deriveFont(Font.BOLD));
+
+		// Cred edit
+		JButton credEdit = new JButton(Messages.getString("NetworkTab.54"));
+		credEdit.addActionListener(new ActionListener() {
+			@Override
+			public void actionPerformed(ActionEvent e) {
+				JPanel tPanel = new JPanel(new BorderLayout());
+				String cPath = (String) PMS.getConfiguration().getCustomProperty("cred.path");
+
+				if (StringUtils.isEmpty(cPath)) {
+					cPath = (String) PMS.getConfiguration().getProfileDirectory() + File.separator + "UMS.cred";
+					PMS.getConfiguration().setCustomProperty("cred.path", cPath);
+				}
+
+				final File cred = new File(cPath);
+				final boolean newFile = !cred.exists();
+				final JTextArea textArea = new JTextArea();
+				textArea.setFont(new Font("Courier", Font.PLAIN, 12));
+				JScrollPane scrollPane = new JScrollPane(textArea);
+				scrollPane.setPreferredSize(new java.awt.Dimension(900, 450));
+
+				if (!newFile) {
+					try {
+						FileInputStream fis = new FileInputStream(cred);
+						BufferedReader in = new BufferedReader(new InputStreamReader(fis));
+						String line;
+						StringBuilder sb = new StringBuilder();
+						while ((line = in.readLine()) != null) {
+							sb.append(line);
+							sb.append("\n");
+						}
+						textArea.setText(sb.toString());
+						fis.close();
+					} catch (Exception e1) {
+						return;
+					}
+				} else {
+					StringBuilder sb = new StringBuilder();
+					sb.append("# Add credentials to the file");
+					sb.append("\n");
+					sb.append("# on the format tag=user,pwd");
+					sb.append("\n");
+					sb.append("# For example:");
+					sb.append("\n");
+					sb.append("# channels.xxx=name,secret");
+					sb.append("\n");
+					textArea.setText(sb.toString());
+				}
+
+				tPanel.add(scrollPane, BorderLayout.NORTH);
+
+				Object[] options = {Messages.getString("LooksFrame.9"), Messages.getString("NetworkTab.45")};
+				if (
+					JOptionPane.showOptionDialog(
+						(JFrame) (SwingUtilities.getWindowAncestor((Component) PMS.get().getFrame())),
+						tPanel,
+						Messages.getString("NetworkTab.54"),
+						JOptionPane.OK_CANCEL_OPTION,
+						JOptionPane.PLAIN_MESSAGE,
+						null,
+						options,
+						null
+					) == JOptionPane.OK_OPTION
+				) {
+					String text = textArea.getText();
+					try {
+						FileOutputStream fos = new FileOutputStream(cred);
+						fos.write(text.getBytes());
+						fos.flush();
+						fos.close();
+						PMS.getConfiguration().reload();
+					} catch (Exception e1) {
+						JOptionPane.showMessageDialog((JFrame) (SwingUtilities.getWindowAncestor((Component) PMS.get().getFrame())), Messages.getString("NetworkTab.55") + e1.toString());
+					}
+				}
+			}
+		});
+		builder.add(credEdit, FormLayoutUtil.flip(cc.xy(1, 3), colSpec, orientation));
 
 		JComponent availablePluginsHeading = builder.addSeparator(Messages.getString("PluginTab.1"), FormLayoutUtil.flip(cc.xyw(1, 5, 9), colSpec, orientation));
 		availablePluginsHeading = (JComponent) availablePluginsHeading.getComponent(0);
